fix(providers-ui): expose Reasoning Rewrite adapter at provider level
The provider adapter pane filtered reasoning_rewrite out of its checkbox
grid because the rules options editor only existed in the per-model
dialog. Extract the rules editor into a shared
ReasoningRewriteRulesEditor component used by both ProviderModelsEditor
and ProviderAdvancedEditor, and show the adapter checkbox (with its
rules editor) in the provider pane. Provider-level reasoning_rewrite
config was previously editable only via API.
